drm/gem: Create a drm_gem_object_init_with_mnt() function
commit 9aa5f00f839d261bfef937425e9c2e929c0db118 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Commit 0992b2541e1cd9580c2e70fab7a78558de054bae upstream

For some applications, such as applications that uses huge pages, we might
want to have a different mountpoint, for which we pass mount flags that
better match our usecase.

Therefore, create a new function `drm_gem_object_init_with_mnt()` that
allow us to define the tmpfs mountpoint where the GEM object will be
created. If this parameter is NULL, then we fallback to `shmem_file_setup()`.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240923141348.2422499-5-mcanal@igalia.com
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>