remoteproc: k3-r5: keep the device on during stop
commit a04282dfaa0ae1917d43cc3f0bbdd8787a4688e6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Keep the core on when stopping a R5F core. On a stop, the core is
halted and held in reset. On a start, the reset is released after a
firmware loaded and this boots the core.
This mechanism allows each R5F core in a cluster to be
independently stopped and loaded with a new firmware.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Reviewed-by: Beleswar Padhi <b-padhi@ti.com>
Tested-by: Beleswar Padhi <b-padhi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>