v4l2-controls: Add Control for Background Detection
commit 4fefde8d64f10d37ffa06307905a73a79a6d1df0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Background detection is a useful technique in video encoding to preserve
bits in cases when the video being encoded has a background that doesn't
change often. A predominant use case for this is security scenarios where
the background hardly changes. Some video IPs offer this feature as part of
their hardware. Expose this control so end-users can select this feature if
offered by the IP.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>