watchdog: cpu5wdt.c: Fix use-after-free bug caused by cpu5wdt_trigger
commit 573601521277119f2e2ba5f28ae6e87fc594f4d4 upstream.

When the cpu5wdt module is removing, the origin code uses del_timer() to
de-activate the timer. If the timer handler is running, del_timer() could
not stop it and will return directly. If the port region is released by
release_region() and then the timer handler cpu5wdt_trigger() calls outb()
to write into the region that is released, the use-after-free bug will
happen.

Change del_timer() to timer_shutdown_sync() in order that the timer handler
could be finished before the port region is released.

Fixes: e09d9c3e9f85 ("watchdog: cpu5wdt.c: add missing del_timer call")
Signed-off-by: Duoming Zhou <duoming@zju.edu.cn>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Link: https://lore.kernel.org/r/20240324140444.119584-1-duoming@zju.edu.cn
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Wim Van Sebroeck <wim@linux-watchdog.org>
[Zhe: Fix for CVE-2024-38630. The function timer_shutdown_sync in the
original fix is not introduced to 5.10 yet. But as stated in f571faf6e443
("timers: Provide timer_shutdown[_sync]()"), "timer_shutdown_sync() has the
same functionality as timer_delete_sync() plus the NULL-ification of the
timer function." So timer_delete_sync is enough for this case.]
Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>