LF-13761-01 arm64: dts: imx93: Set default frequency for video pll
commit  a75b208d234e8d1e264905ec43adfafcaf68f5b5 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

After SoC PoR, The video PLL's frequency setting may not be in the
correct design frequency range & out of spec. According to the
SoC Datasheet, the max output frequency of the video PLL is 594MHz,
so set the default frequency to that value for this PLL in clock
driver init stage.

When this video PLL is used as clock source by any device for any
purpose, User can NOT rely on this default frequency of video PLL,
and must set this video PLL frequency explicitly based on the use case.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Tested-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Tested-by: Ying Liu <victor.liu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>