regulator: tps6594-regulator: Fix the number of irqs for TPS65224 and TPS6594
commit 6fd6965e35786e1cc915171f2ff70b840d981ca9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The number of irqs is computed to allocate the right amount of memory for
the irq data. An array of struct tps6594_regulator_irq_data is allocated
one time for all the irqs. Each irq uses one cell of the array.

If the computed number of irqs is not correct, not allocated memory could
be used.

Fix the	values used in the calculation for TPS6594 and TPS65224.

Fixes: 00c826525fba ("regulator: tps6594-regulator: Add TI TPS65224 PMIC regulators")
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>