octeontx2-af: fix to get different rq mask
commit 34a9cd9b9b19df92c4ed75d51b66a304b30ac463 from
git@git.assembla.com:cavium/WindRiver.linux.git

Currently we are getting same RQ mask id for different profile
as RQ mask id calculation is based on logical AND operation
between mask and set. We need to get different RQ mask for
different profile if there is a change in either mask or set
or both. This patch fixes the same.

Fixes: 8bdf0536aa82 ("octeontx2-af: add support for CPT second pass")

Change-Id: I4a7c66bb0edb9b8b417188377ab3702e2b09419b
Signed-off-by: Rakesh Kudurumalla <rkudurumalla@marvell.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>