usb: dwc3: core: add support for setting NAK enhancement bits for FS/HS
commit  fd0990fcb52a54031ec16f904315c50c953126fa from
https://github.com/raspberrypi/linux.git rpi-6.18.y

If a device frequently NAKs, it can exhaust the scheduled handshakes in
a frame. It will then not get polled by the controller until the next
frame interval. This is most noticeable on FS devices as the controller
schedules a small set of transactions only once per full-speed frame.

Setting the ENH_PER_NAK_FS/LS bits in the GUCTL1 register increases the
number of transactions that can be scheduled to Async (Control/Bulk)
endpoints in the respective frame time. In the FS case, this only
applies to FS devices directly connected to root ports.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>