MLK-26173: dmaengine: fsl-edma-v3: fix dmatest memcpy fails on i.MX 8QM
commit 215f33ce24fd38e9d860333a46c68496a6c3b127 from
https://github.com/nxp-imx/linux-imx.git

The dmatest memcpy test is MEM_TO_MEM that means read and write memory.
The edma access the memory that is remote according to the imx8qm edma ADD.
The dmatest memcpy test fail because the access is configured as local.

The test fail log as follow:
[ 66.268506] dmatest: dma0chan0-copy0: result #1: 'test timed out' with src_off=0x100 dst_off=0x80 len=0x3ec0 (0)
[ 66.278785] dmatest: dma0chan0-copy0: summary 1 tests, 1 failures 0.32 iops 4 KB/s (0)

This patch change local into remote config.

This patch has potential problems in follow special case.
1.request edma chan.
2.prepare_slave_sg description
3.submit description
4.prepare_memcpy
5.submit description
6.prepare_slave_sg description
7.submit description
8.issue_pending_work

If the is_remote flag of the first and second prepare_slave_sg is flase.
But the is_remote flag of the prepare_memcpy is true. Finally, the
is_remote flag is changed to be true. It will cause the first and second
prepare_slave_sg transfer fail. we don't find better solution at present.

In addition, we current cases are simple, we don't meet such case.

Reviewed-by: Frank Li <Frank.Li@nxp.com>
Signed-off-by: Joy Zou <joy.zou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>