HSD #15011993515-3: firmware: stratix10-svc: Add support for asynchronous communication with Stratix 10 service channel
commit 8f00455c92a1702582423a609f00457aeb8c955b from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

This commit adds support for asynchronous
with the Stratix 10 service channel. It introduces
new functions and definitions to enable asynchronous
messaging and handling of callbacks. The changes
include the addition of asynchronous client registration,
sending asynchronous messages, polling the status of
asynchronous requests, and completing
transactions.

The new functions added are:
- stratix10_svc_add_async_client: Adds an client
        to the service channel.
- stratix10_svc_remove_async_client: Removes an
        asynchronous client from the service channel.
- stratix10_svc_async_send: Sends an asynchronous
        message to the SDM mailbox in EL3 secure firmware.
- stratix10_svc_async_poll: Polls the status of an
        asynchronous service request.
- stratix10_svc_async_done: Completes an
        asynchronous transaction.

These changes enhance the functionality of the
Stratix 10 service channel by allowing for more
efficient and flexible communication with the firmware.

Signed-off-by: Mahesh Rao <mahesh.rao@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>