net: hsr: Add VLAN CTAG filter support
commit 2f6651c0bea4ed1cbb85fd41da56ba90fefb44ae from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch adds support for VLAN ctag based filtering at slave devices.
The slave ethernet device may be capable of filtering ethernet packets
based on VLAN ID. This requires that when the VLAN interface is created
over an HSR/PRP interface, it passes the VID information to the
associated slave ethernet devices so that it updates the hardware
filters to filter ethernet frames based on VID. This patch adds the
required functions to propagate the vid information to the slave
devices.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>