net: ti: icssg-prueth: Add VLAN support for HSR mode
commit 130fac47786053860dc667ea9ba42a13ed013c77 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support for VLAN addition/deletion in HSR mode.
In HSR mode, even if the host port is not a member of
the VLAN domain, the slave ports should simply forward the
frames. So allow forwarding of all VLAN frames in HSR mode.

Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>