arm64: dts: ti: k3-am642-evm/sk: Drop routing GenF to SYNC1_OUT signal
commit a95cc1059ec665856ef3e17d99dc1c3671aaa80e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

GenF was routed to SYNC1_OUT with the intent of having a way to measure
the PPS signal, since the board schematics indicated that SYNC1_OUT_TP
was connected to the J12 header. However, the datasheet indicates that
it is actually SYNC0_OUT and not SYNC1_OUT that is connected to the D18
pin which is routed to the J12 header. Hence, don't route GenF to
SYNC1_OUT since it doesn't serve the original purpose of providing a way
to measure the PPS signal.

Fixes: 18645b81a4be ("arm64: dts: ti: k3-am642-evm/sk: Add CPSW3G CPTS PPS support")
Signed-off-by: Meghana Malladi <m-malladi@ti.com>
Reviewed-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>