media: i2c: imx477: Add options for slightly modifying the link freq
commit f22cbab9d74da4c60c989cdb15d037a9264094d7 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The default link frequency of 450MHz has been noted to interfere
with GPS if they are in close proximty.
Add the option for 453 and 456MHz to move the signal slightly out
of the band. (447MHz can not be offered as corruption is then observed
on the 133x992 10bit mode).

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

fixup imx477 gps

Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>