TEMP: media: imagination: vxe-vxd: encoder: Fix RGB Crash
commit 541c20281af79a7df96bb94b4e3a923092d7ceff from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

IMG VXE Encoder supports reading in raw RGB videos and converting them
to be encoded in NV12 format. The colorspace converter has a requirement
where the width must be aligned to 64. In cases where the width is not,
alignment fails and colorspace converter hangs.

A deeper dive is needed into the calculation of the coefficients used to
program the colorspace converter. In the meantime, lets reject invalid
widths so crashes do not occur.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Acked-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>