net: enetc: Do check the number of BDs in XDP_TX
When we were porting the SDK commit d4defbd2faa9 ("LF-12807-5 net: enetc:
check the number of BDs needed for XDP_TX frame") to linux-yocto, we
missed some codes due to the merge conflicts. Add the check of BD number
back as what the original SDK commit want to do.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>