media: platform: vxe-vxd: Fix the variable shadowing issue in macro
The variable __value is declared in both REGIO_WRITE_FIELD and
REGIO_WRITE_FIELD_LITE macros, both of them also invoke
REGIO_CHECK_VALUE_FITS_WITHIN_FIELD macro and pass the __value as a
parameter for it. But the __value is also declared in macro
REGIO_CHECK_VALUE_FITS_WITHIN_FIELD, so this will produce incorrect
result and cause the following build error:
  In file included from drivers/media/platform/vxe-vxd/decoder/vxd_int.c:27:
  drivers/media/platform/vxe-vxd/decoder/vxd_int.c: In function ‘vxd_set_reconpictcmds’:
  drivers/media/platform/vxe-vxd/decoder/reg_io2.h:36:22: error: ‘__value’ is used uninitialized [-Werror=uninitialized]
     36 |                 type __value = value; \
        |                      ^~~~~~~
  drivers/media/platform/vxe-vxd/decoder/reg_io2.h:64:9: note: in expansion of macro ‘REGIO_CHECK_VALUE_FITS_WITHIN_FIELD’
  ITHIN_FIELD(group, reg, field, __value, type); \
        |         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  drivers/media/platform/vxe-vxd/decoder/vxd_int.c:552:9: note: in expansion of macro ‘REGIO_WRITE_FIELD_LITE’
  ict_cmds[VDECFW_CMD_CODED_PICTURE],
        |         ^~~~~~~~~~~~~~~~~~~~~~

Add an addition '_' for the variable in macro
REGIO_CHECK_VALUE_FITS_WITHIN_FIELD to fix this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>