octeontx2-pf: Add correct page pool parameters
commit 61dd5be8d606f56077485cf494ee587b0f7e5568 from
git@git.assembla.com:cavium/WindRiver.linux.git

The change to use correct page pool parameters are part of CN20K patch.
Since as part of this 6.1 to 6.6 porting, we are not porting the cn20k
changes, this patch is a temporary fix to enable PF/VF and workings of
networkings.

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I1378ba527b8768af988a1a811bebdbe56d3bd042
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/139726
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>