MGS-7950 [#imx-3361] Coverity: fix free of array-typed value
commit  060bf647556d15a20c140f6502839372d0246f49 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

According to the Coverity report, the second param can't be a array type
in thermal_of_cooling_device_register, otherwise there will be a free
of array-typed value issue in the failing path in this function.
We use a pointer instead as a fix.

Signed-off-by: Guangliu Ding <guangliu.ding@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>