net: enetc: Fix dependency issue for the enetc ptp clock driver
The enetc ptp clock driver optionally depends on either FSL_ENETC or
FSL_ENETC_VF, as specified by the following expression:
	FSL_ENETC || FSL_ENETC_VF

Howerver, this condition is insufficient. We must also ensure that
the ptp driver can't be built as y if either FSL_ENETC or FSL_ENETC_VF
is built as m. So additional dependencies are added to resolve this
issue.

Reported-by: Wang Mingyu <wangmy@fujitsu.com>
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>