mtd: spi-nor: Store GPIO info for resetting device during resume
commit 103a9f2a2c998aacf45e43e2c06c6443e60e2f4d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Upon resume, SPI-NOR triggers the hardware reset API to reset the flash
device by toggling the GPIO. This step is necessary because after resume,
the flash needs to be put in SDR mode. To facilitate this, SPI-NOR needs
GPIO information to reset the flash device. Therefore, a gpio_desc member
is added to the spi_nor structure to retain the GPIO information and
utilize it for resetting the device.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>