LF-14152-4 mfd: max96789-core: power cycle the chip using gpio
commit  626903d475b2e7c31e8230638aaa9ed07a843c58 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

If the platform is soft rebooted, then the chip is not reset and the
drivers will fail to probe because GMSL links are left with the
previous settings.

For this implementation to work properly, the serializer board needs a
HW workaround: resistors R81 and R82 need to be removed.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>