net: dpaa2-switch: initialize the status variable with zero
commit  969b0617c9103b7f9a956274dba293dc22749832 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

In the blamed commit we tried to not clear any interrupts automatically
through the dpsw_get_irq_status() command. We did this by not having the
status parameter initialized with all ones. The problem is that we did
not initialize the variables with all zeroes, so now we are just
hoping that the variable is indeed zero.
Make sure that status is zero and initialize the variable.

Fixes: f6da276479c6 ("dpaa2-switch: do not clear any interrupts automatically")
Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>