LF-13948-5: media: imx8mq-mipi-csi2: fix get_fmt with streams
commit  83ea6b06f5ae94d8426e542a5791b43938d933a3 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Currently, the get_fmt does note care about the stream received on the
format, so it only set the format on stream 0. Fix this in order to be
able to set the format on all the exposed streams.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>