FROMLIST: media: ti: j721e-csi2rx: allocate DMA channel based on context index
commit d59fa7f751e131f7f1daf7cf8464dbaee5ca3261 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

With multiple contexts, there needs to be a different DMA channel for
each context. Earlier, the DMA channel name was hard coded to "rx0" for
the sake of simplicity. Generate the DMA channel name based on its index
and get the channel corresponding to the context.

Link: https://lore.kernel.org/all/20240627-multistream-v2-5-6ae96c54c1c3@ti.com/

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Yemike Abhilash Chandra <y-abhilashchandra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>