FROMLIST: media: ti: j721e-csi2rx: add a subdev for the core device
commit a8154e143d2c85c1db60db7a01c51e092052be91 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

With single stream capture, it was simpler to use the video device as
the media entity representing the main TI CSI2RX device. Now with multi
stream capture coming into the picture, the model has shifted to each
video device having a link to the main device's subdev. The routing
would then be set on this subdev.

Add this subdev, link each context to this subdev's entity and link the
subdev's entity to the source. Also add an array of media pads. It will
have one sink pad and source pads equal to the number of contexts.

Link: https://lore.kernel.org/all/20240627-multistream-v2-6-6ae96c54c1c3@ti.com/

Co-developed-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Yemike Abhilash Chandra <y-abhilashchandra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>