FROMLIST: media: ti: j721e-csi2rx: add multistream support
commit 7d2a51bee9113388bcc2d0927153f392d83c7709 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Each CSI2 stream can be multiplexed into 4 independent streams, each
identified by its virtual channel number. To capture this multiplexed
stream, the application needs to tell the driver how it wants to route
the data. It needs to specify which context should process which stream.
This is done via the new routing APIs.

Add ioctls to accept routing information from the application and save
that in the driver. This can be used when starting streaming on a
context to determine which route and consequently which virtual channel
it should process.

Support the new enable_stream()/disable_stream() APIs in the subdev
instead of s_stream() hook. We wait for the userspace to start capturing
on all video nodes that have active routes, and once all video nodes are
STREAMON, we request the source to enable_stream() for all the sink
streams.

Link: https://lore.kernel.org/all/20240627-multistream-v2-12-6ae96c54c1c3@ti.com/

Co-developed-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Yemike Abhilash Chandra <y-abhilashchandra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>