Revert "arm: dts: imx6q-sabresd: Enable imx6q pcie"
This reverts commit ffb7412ca5fbc3267803eff102a40fef061ee9f8

There is a separate DTS file specifically responsible for enabling
PCIe, so PCIe does not need to be enabled in the imx6q-sabresd.dts.

In fact, using the imx6q-sabresd-pcie.dtb provided by the SDK also
fails to properly enable the PCIe feature. This is an issue caused
by the SDK. In the mainline kernel, PCIe functionality works as
expected. We haven't resolved this issue yet, so for now, we have
to stay in sync with the SDK.

Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>