spi: cadence-quadspi: Retrieve all chip-select values from the flash node
commit cae9073dbbcfb1a4f3e9fcab1228669c358c7263 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

In stacked mode, the driver does not account for the second chip-select
value when setting the ->num_chipselect field in the struct cqspi_st and
struct spi_controller to the current chip-select count in commit
0f3841a5e115 ("spi: cadence-qspi: report correct number of chip-select"),
leading to a probe failure. Update the driver to retrieve all chip-select
values from the flash node.

Fixes: 0f3841a5e115 ("spi: cadence-qspi: report correct number of chip-select")
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Message-ID: <20250101183721.2869197-1-amit.kumar-mahapatra@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>