net: axienet: Fix lost wake-up scenario in xmit
commit fc71e47023b40bb8c792002f8274c756adcfd2ab from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

In ndo_start_xmit we are checking for free BD count
and incase there are no available BD's we stop
the networking stack.

If these sequence of operation are preemted then
it will lead to classic Lost Wake-Up Problem.

If interrupt comes after xmit has checked  BD
count and TX post processing clears up the BD and
issue wake-up and then later xmit goes for a wait.
There is no one to wake up xmit and as wake-up
event is lost.

Signed-off-by: Appana Durga Kedareswara Rao <appana.durga.rao@xilinx.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Message-ID: <20250107104507.359136-2-suraj.gupta2@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>