net: axienet: Add support for 10G/25G MAC
commit 3669bd1a79776ca27ad0215551d665bea7999b8a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add support for 10G/25G MAC.
>From IP version 3.2, XXV Ethernet IP has a new GT reset status
register which provides information on whether TX and RX clock
is stable. There is a limitation in the IP where RX Block Lock
Status register cannot be read if GT RX recovered clock is not
ready. Hence this new register needs to be polled for GT reset
complete and then the Block Lock should be polled for link up.
This new register should be polled after GT reset is done and
before any other register access.

Signed-off-by: Appana Durga Kedareswara Rao <appana.durga.rao@xilinx.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Message-ID: <20250107104507.359136-10-suraj.gupta2@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>