net: xilinx: Fix phylink_of_phy_connect error on RGMII designs
commit 87361f292ef3b94a0238e3aeaf3a6b35ff202883 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

For PCS/PMA based designs where PHY mode can be sgmii or 1000BaseX,
the existing implementation checks for pcs-handle and phy-handle
properties. For other modes (e.g RGMII), the implementation does
not check for these properties which results in PHY failures.

Update the code to ensure that pcs-handle and phy-handle properties
are parsed for all cases except XXV and MRMAC which do not have PHY
in the designs.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Message-ID: <20250107104507.359136-19-suraj.gupta2@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>