dt-bindings: net: xlnx,axi-ethernet: Add bindings documentation for Ethernet Offload Engine
commit 67303eac15c82f3aece9f872be4717829cdd56bd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add devicetree bindings documentation for Ethernet Offload Engine.

The ethernet offload engine support features offloaded from software to
the hardware to increase the performance and efficiency of the system.
These offloaded features include calculating checksum and segmentation
offloading of the packet if its size is greater than MTU and in Rx, it
supports the Receive Offload feature.

It is an independent IP and can be integrated into existing ethernet
subsystem IPs.

Co-developed-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Ravikanth Tuniki <ravikanth.tuniki@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Message-ID: <20250107104507.359136-24-suraj.gupta2@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>