net: axienet: Add EOE GSO and checksum support for TX
commit 16e693b85e01aac87c7cfcf4cf3ce7611367622a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Generic Segmentation Offload performs the segmentation of the
UDP packet i.e. creating multiple small packets of 1472 bytes
from a larger packet of up to 64KB size. Tx coalesce count is
set to 1. With offload enabled, there are not as many interrupts
as before and the interrupt for every 64KB segment needs to be
handled immediately to ensure better performance.

Signed-off-by: Kiran kumar Gandham <kirankumar.gandham@amd.com>
Signed-off-by: Ravikanth Tuniki <ravikanth.tuniki@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Message-ID: <20250107104507.359136-26-suraj.gupta2@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>