net: dpaa: integrate SerDes phys with lynx pcs
commit  131c674f699fca4d1c7bd37a050abd1565502b04 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The lynx pcs gained support for managing SerDes lanes on its own, so
let's make dpaa-eth use that as well.

Note that the hardware on which this driver runs is integrated with the
Lynx 10G SerDes, for which a Linux driver doesn't exist. So in mainline,
the existing SerDes lane handling in dpaa-eth is dummy code.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>