LF-15623-7 net: dsa: netc: add basic suspend and resume support
commit  ab9ad71929fde7f06ea29ae7b298e3edcf5c57ee from
https://github.com/nxp-imx/linux-imx lf-6.18.y

There are two cases of NETC when the system enters the suspend mode, one
is powered off, the other is still powered on. Currently we only know
that NETC won't be powered off in suspend mode when ENETC WOL is enabled.
For other cases, there is no helper function for us to query whether NETC
is powered off in suspend mode. So we assume that NETC is powered off,
and all hardware configurations will be lost in suspend mode. Therefore,
the driver needs to restore the configurations when the switch resumes.

For the case that NETC is powered off in suspend mode, currently, we have
completed the recovery of registers, VLAN filter table, egress treatment
table and egress counter table, FDB table, and PTP related IPFT entries.
However, there are still some tables that have not been restored, such as
time gate scheduling table and tc-flower related tables, such as ingress
stream sdentification table, ingress stream table, stream gate instance
table, etc.. The recovery of these tables will be supported in subsequent
patches.

For the case that NETC is still powered on in suspend mode (ENETC WOL is
enabled), it is relatively simple, because all hardware configurations
will not be lost, so we only need to disable the switch PCIe device when
the system suspends, disable the related clocks and work queue, and then
re-enable these functions when the system resumes.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>