arm64: zynqmp: Describe ethernet controllers via aliases on SOM
commit 78d8f65d7c47a00b6e3aff770a3ae1e18f0f34ca from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add ethernet aliases to CC (Carrier card) description to create a
connection which is used by fdt_fixup_ethernet() for updating
local-mac-address in DT.
On Kria SOM MAC address is read from i2c eeprom at start and based on it
environment variables are created. Without creating aliases U-Boot is not
able to inject local-mac-address DT property and OS won't get the same MAC
address unless another i2c read is happening in OS.
Also aliases are using string not phandle that's why full path has to be
provided but that shouldn't be a big issue because location of ethernet
controller is fixed.

Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>