PENDING: net: ethernet: ti: add RPMsg structures for Ethernet Switch Firmware
commit c1676b296d0475d43d90c84b6c3dc295c3a39f1e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The Ethernet Switch Firmware (EthFw) running on the remote MAIN R5 core is
capable of configuring the CPSWnG instance of TI's CPSW Ethernet Switch
on various TI SoCs such as J721E, J7200 and J784S4. EthFw is in complete
control of CPSWnG and acts as a server which allocates and provides the
resources to clients on different cores. All communication between EthFw
and the clients occurs via RPMsg.

Define the message structures used to communicate with EthFw. This shall
enable the Linux Client to avail the services provided by EthFw.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>