PENDING: net: ethernet: ti: cpsw-proxy-client: enable message exchange with EthFw
commit 2ed69f3cfa8a611ce3d7a3228e81ffd561d216b9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a new function which sends RPMsg requests to EthFw and shares the
response for the request. The RPMsg callback function copies the response
it receives from EthFw to the driver's private member, thereby allowing
the response to be shared with the newly added function which sends the
request.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>