PENDING: net: ethernet: ti: cpsw-proxy-client: enable client driver functionality
commit 16d8434d3b5be9c8d29b5dd6aaedb7e249ca3362 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Use the helpers added so far to enable the client driver functionality.

Depending on the SoC, the type of DMA and therefore the compatible
corresponding to it vary. Map the DMA compatible to the SoC in order to
support SoCs with different types of DMA.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>