PENDING: net: ethernet: ti: cpsw-proxy-client: enable reset-recovery support
commit d5bf785f5cfb2f54049c859870f5bd47c998dcbe from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In the event of Hardware Error, EthFw is capable of detecting the Error
and notifying all Clients of this error. Once the Clients handle the
Error notification from EthFw and notify the same to EthFw, EthFw begins
the recovery sequence for CPSW. Using the existing helpers for handling
the error and recovery completion notifications from EthFw, enable the
Client driver support for reset-recovery sequence.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>