FROMLIST: arm64: dts: ti: k3-j784s4-j742s2-evm: Add overlay to enable USB0 Type-A
commit b7ca80adb0303da5a2f7aef539aef5a9fc7f9d2d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The USB0 instance of the USB controller on both the J742S2 EVM and the
J784S4 EVM supports a single USB interface at a time among the following:
1. USB3.1 Gen1 Type C interface
2. Two USB2.0 Type A interfaces via an on-board USB Hub.

By default, the USB3.1 Gen1 Type C interface is supported on both of the
EVMs. Enable the USB2.0 Type A interface by configuring the USB2.0_MUX_SEL
mux. Additionally, set the Dual-Role Mode to Host since a Type-A interface
is only associated with the Host Mode of operation.

Link: https://patch.msgid.link/20250124124150.2024963-1-s-vadapalli@ti.com
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>