s32cc: pcie: Simplify PCIe link up procedure
commit 064204b44fc56784f9daafa89e19c5250f680c2e from
https://github.com/nxp-auto-linux/linux

PCIe RC linkup procedure was more complex than necessary.
Previously, it followed below sequence:
1. disable LTSSM
2. Set Speed to Gen1
3. enable LTSSM
4. Wait for link up
5. Set higher speed modes (Gen2 or Gen3)
6. Initialize speed change by setting GEN2_CTRL_OFF.DIRECT_SPEED_CHANGE

New approach is to simply let LTSSM negociate the maximum speed rate, which
should be GEN3 for PCIe S32CC controller.

If because of certain hardware EndPoint setups, the maximum speed rate
needs to be limited, then 'max-link-speed' entry via DTS can be used to
limit either GEN1 or GEN2 PCIe speeds. Setting it to GEN3 is not and
should not be mandatory (as it previously was). If set to GEN3 it will
not take any effect.

Issue: ALB-12513
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>