xilinx-ai-engine: Add tracepoint support
commit 573298097ac29974170ca66421120d853dd457d7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add tracepoints to partition and aperture IOCTls. Tracepoints were also
added to interrupt backtracking and register write.

This for internal development and debugging purpose to better understand
and debug the flow and register writes/reads driver performs.

These operations are NOP when TRACING is not enabled in kernel config.

Signed-off-by: Govindarajulu Varadarajan <govindarajulu.varadarajan@amd.com>
Message-ID: <20250128022005.205882-3-govindarajulu.varadarajan@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>