v4l: xilinx: sdirxss: Update the device-tree flags
commit f86139574b755ae0a6cc3fd756a8701d64541025 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The properties reset_gt and picxo_reset have been updated to reset-gt and
picxo-reset respectively, to comply with the device tree binding
conventions.

Informational messages are added to notify users when deprecated properties
are detected. The code around the deprecated property names will be removed
in a future upgrade.

Signed-off-by: Dhanunjanrao Katta <katta.dhanunjanrao@amd.com>
Message-ID: <1738144255-12694-3-git-send-email-katta.dhanunjanrao@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>