FROMLIST: rpmsg: char: Export alias for RPMSG ID rpmsg-raw from table
commit bb0033a76cdd2b82426fd452fedfddf2a8208579 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Module aliases are used by userspace to identify the correct module to
load for a detected hardware. The currently supported RPMSG device IDs for
this module include "rpmsg-raw", but the module alias is "rpmsg_chrdev".

Use the helper macro MODULE_DEVICE_TABLE(rpmsg) to export the correct
supported IDs. And while here, to keep backwards compatibility we also add
the other ID "rpmsg_chrdev" so that it is also still exported as an alias.

This has the side benefit of adding support for some legacy firmware
which still uses the original "rpmsg_chrdev" ID. This was the ID used for
this driver before it was upstreamed (as reflected by the module alias).

Link: https://lore.kernel.org/all/20240729164527.340590-1-afd@ti.com
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>