FROMLIST: remoteproc: k3: Refactor rproc_reset() implementation into common driver
commit d50ef190d7e8873cc3773cbd446dd33a8368862e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The k3_dsp_rproc_reset() function erroneously asserts the local reset
even for devices which do not support it. Even though it results in a
no-operation, Update the logic to explicitly assert the local reset for
devices that support it and only the global reset for those that do not.

Further, refactor the above function into the ti_k3_common.c driver as
k3_rproc_reset() and use it throughout DSP and M4 drivers for resetting
the remote processor.

Link: https://lore.kernel.org/all/20250103101231.1508151-7-b-padhi@ti.com/
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>