PENDING: v4l2-controls: Add Control for Background Detection
commit 7fb86761424dd3eb7e1baf7b4cb64f4d5c40d73d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Background detection is a useful technique in video encoding to preserve
bits in cases when the video being encoded has a background that doesn't
change often. A predominant use case for this is security scenarios where
the background hardly changes. Some video IPs offer this feature as part of
their hardware. Expose this control so end-users can select this feature if
offered by the IP.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>