PENDING: media: chips-media: wave5: Add Support for Background Detection
commit 9bb0fb195bd8ba1fed6cf8621dc75ea1d3e077cc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Wave5 encoder can be configured to detect a background region in a frame.
If a background region is detected, it will use less bits or skip mode to
encode that region. This can assist in lowering the video bitrate for a
given stream.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Acked-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>