net: enetc: add ingress stream identifier table support
commit  7f5efb0c29d11a8376511ba7b8fe264e1ae5dcbc from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The Ingress Stream Identifier table contains a set of entries each
capable of classifying incoming traffic using a mix of L2 parsed and
arbitrary data fields, into streams. As a result of an entry match,
an Ingress Stream Entry ID is provided, which is a locally significant
identifier (global to the switch), that identifies a stream within an
ENETC instance or the switch.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>