net: enetc: convert mailbox messages to new formats
commit  f9581d12f4e81fcb5ed8e6e4f1fa6daeca576fee from
https://github.com/nxp-imx/linux-imx lf-6.18.y

On the LS1028A platform, the mailbox between ENETC PF and VF is only used
to change the MAC address of VF. The message is too simple and difficult
to expand, and the receiver cannot verify the integrity of the data, so
a new message format is redefined. The current patch only converts the
original message, and more new message support will be added later.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>