PENDING: drm/bridge: sii902x: Add device_link dependency with drm device
commit 0c14e18358356712fb440883bfd9d727d4c3845a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Create a supplier - consumer relationship between sii902x and the
display-controller device.

This is required to properly support system suspend/resume. The DRM
framework initates a mode_set throughout the bridge chain as part of
display-controller resume. If bridge has not resumed before this, the
mode_setting register-writes on the bridge make it end up in a bad
state.

The sii902x device acts as a supplier, and the display-controller as
consumer. This ensures that the display-controller resume (and by
extension the bridge mode_set) happen after the sii902x has resumed.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>