FROMLIST: arm64: dts: ti: k3-am62p5-sk: Enable IPC with remote processors
commit a75cbd3626b8085513288822666c18859b23f299 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

For each remote proc, reserve memory for IPC and bind the mailbox
assignments. Two memory regions are reserved for each remote processor.
The first region of 1MB of memory is used for Vring shared buffers
and the second region is used as external memory to the remote processor
for the resource table and for tracebuffer allocations.

Link: https://lore.kernel.org/linux-devicetree/20250206235200.3128163-1-jm@ti.com/
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>