PENDING: spi: spi-mem: Tell controller when device is ready for calibration
commit 35a56fd92209b64962f061d0c4e2e5ffe0f5fee7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Some controllers like the Cadence OSPI controller need to perform a
calibration sequence to operate at high clock speeds. This calibration
should happen after the flash is fully initialized otherwise the
calibration might happen in a different SPI mode from the one the flash
is finally set to. Add a hook that can be used to tell the controller
when the flash is ready for calibration. Whether calibration is needed
depends on the controller.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Santhosh Kumar K <s-k6@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>