pinctrl: scmi: Fix build failure by adjusting arguments of of_property_for_each_u32()
This patch is to fix the following build failure:
    .../.../pinctrl-scmi.c:1009:63: error: macro "of_property_for_each_u32" passed 5 arguments, but takes just 3
         of_property_for_each_u32(np, "pinmux", prop, p, pinmux) {

The reason is that, the commit 914ef7d1a702 ("of: remove internal arguments
from of_property_for_each_u32()") drops two parameters, which are primarily
meant as internal variables for the macro.

So updating the parameters to keep align with above change, can fix the issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>