printk: rt: Update con->printk_seq for suppressed message printing
On the RT kernel, we have observed that some stale messages (which were
previously suppressed due to their low log level) are output to the
console when an MCE exception is injected.
# mce-inject <<!
CPU 0 BANK 1
STATUS UNCORRECTED
ADDR 0x1234
!

This issue arises due to the implementation of printk in the RT kernel.
The console printing operates in two modes:
- Asynchronous Mode: Under normal conditions, the console printing is
  fully asynchronous and handled by a kthread.
- Synchronous Mode: In special cases, such as oops, the kernel enters
  into sync mode where console printing is performed using the
  write_atomic() callback.

In the console printing kthread, the con->printk_seq is not updated
immediately if a message is suppressed due to a lower log level. Indeed,
a local seq is used to track the currently printed sequence. When a MCE
exception occurs, the MCE exception handler runs with irq disabled, and
the kernel enters to oops state immediately. As a result, the console
printk kthread can't be woken up by printk() calls. Simultaneously,
the kernel oops invokes bust_spinlocks() and raises the console log level
via console_verbose(). Consequently, the previously suppressed messages
in the printk ring buffer are output to console because the con->printk_seq
has not been updated.

While this may not seem a critical issue at first glance, if the number
of suppressed messages is large (e.g., thousands of lines), it can take
a considerable amount of time to output all these messages. This can be
annoying and could lead to unexpected problems, such as triggering a
watchdog reset and interfering with the expected kdump process.

Fix this issue by updating the con->printk_seq for suppressed message
printing.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>