staging: xilinx-tsn: Fix device tree node memory leak in tsn_ex_ep_probe
commit e785fa9dea12be2c45d698fabeb280849c4e16c9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Fixes a memory leak that occurs when the register_netdev() function fails
in the tsn_ex_ep_probe() function.
The patch introduces a new label free_of_node which is jumped to
when register_netdev() fails. This label ensures that of_node_put(ep_node)
is called to decrement the reference count for the device tree node before
the function returns, preventing the memory leak.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Nagadheeraj Rottela <nagadheeraj.rottela@amd.com>
Message-ID: <20250220090915.1866104-3-srinivas.neeli@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>