LF-14663-2 net: enetc: add ptp timer selection support
commit  8a1c720282996fab052dfe79ad33f880418be128 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

imx95 has only one timer, so a fixed timer configuration is used in
the driver.
On imx94, there are a total of 3 timers. Therefore, it is necessary
to determine which timer to use based on the configuration of dts.

Add nxp,ptp-timer property to decide which timer to use.
If this property does not exist, the default configuration will be
used. The default configuration may be not usable if the timer is
not enabled in dts.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
(cherry picked from commit 95b1dda89b3055915baa867a858adc62841136c2)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>