FROMLIST: mux: mmio: Extend mmio-mux driver to configure mux with new DT property
commit 1f743364a983c894d557096d07295bed7ba6bb47 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

MMIO mux driver is designed to parse "mux-reg-masks" and "idle-states"
property independently to configure mux registers. Drawback of this
approach is, while configuring mux-controller one need to specify every
register of memory space with offset and mask in "mux-reg-masks" and
register state to "idle-states", that would be more complex for devices
with large memory space.

Add support to extend the mmio mux driver to configure a specific register
or set of register in memory space.

Link: https://lore.kernel.org/r/20250304102306.2977836-3-c-vankar@ti.com/
Reviewed-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Chintan Vankar <c-vankar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>