PENDING: arm64: dts: ti: k3-am642-evm/sk: Drop routing GenF to SYNC1_OUT signal
commit 869688c7ab2c7daa597cfbdf3e51da09355070a6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

GenF was routed to SYNC1_OUT with the intent of having a way to measure
the PPS signal, since the board schematics indicated that SYNC1_OUT_TP
was connected to the J12 header. However, the datasheet indicates that
it is actually SYNC0_OUT and not SYNC1_OUT that is connected to the D18
pin which is routed to the J12 header. Hence, don't route GenF to
SYNC1_OUT since it doesn't serve the original purpose of providing a way
to measure the PPS signal.

Signed-off-by: Meghana Malladi <m-malladi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>