dts: s32gxxxardb: Add INA231 sensor node on I2C4 bus
commit d61957cf757911bcb43c2ba6c17e4bc657f5fbe3 from
https://github.com/nxp-auto-linux/linux

The INA231 sensor is available only on some RDB2 revisions, so it must
be disabled by default to ensure compatibility across all boards. This
commit adds the INA231 device node, with status 'disabled' and it will
be enabled at runtime via U-Boot on the supported revisions of RDB2
boards.

Issue: ALB-11752
Signed-off-by: Georgiana Ortila <georgiana.ortila@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>