PENDING: media: ti: j721e-csi2rx: Use pm_notifier instead of suspend/resume hooks
commit 7b5a02b502c0d14abe4c07992edaa46b7442b97f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

We need to stop all on-going streams before system suspend and enable
them back when the system wakes up from sleep. Using .suspend/.resume
callbacks is not optimal, as the order of those callbacks amongst
various devices in the camera pipeline (like sensor, CSI bridge etc.) is
hard to enforce, and can lead to situations where all dependencies are
not yet woken up when we request to start streaming again in the .resume
hook.

Thus, we register a notifier with the PM framework that gets triggered
with events when the system is fully functional, i.e. before any drivers
are suspended, and after all of them have resumed, where we can cleanly
stop or start the streams.

Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>
[Ported to ti-linux-6.12.y-cicd]
Signed-off-by: Rishikesh Donadkar <r-donadkar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>