TI: arm64: dts: ti: k3-am62l3-evm-lpmdemo: Enable support for USB wakeup
commit a50f88d15373e4dbcbd6a8a85113af221ec686f0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The AM62Lx SoC has two instances of USB namely USB0 and USB1, both of
which are brought out on the AM62L3-EVM, with USB0 connected to the
Type-C interface on the EVM and USB1 connected to the Type-A interface.

USB events on any of USB0 and USB1 may be used to wakeup the SoC after
it enters the Deep Sleep  state. Hence, add support for it.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Acked-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>