PENDING: drivers: crypto: sa2ul: Add more checks before processing ipad/opad
commit dd349f9ae191494e03eeec6eaba994607a5359e7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When the key_sz is greater than the block_size then we need to hash it
with the transform to process it further. To avoid repeated hashing for
the key for individual ipad/opad, merge the prepare_kipad and
prepare_kopad in sa_prepare_iopad itself.

Also as per fips standard, there is a minimum key size requirement so
check that based on fips enablement to return the appropriate error
code.

Link: https://datatracker.ietf.org/doc/html/rfc2104
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>