PENDING: media: i2c: ds90ub960: Support multi-channel sensors
commit 94d1948c80891e2535b0423fa72e94c3b5602570 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Before this, the deserializer would ignore the virtual channel (VC) ID
set in the incoming CSI packets, and set the ID to the RX port number in
the outgoing packets. This was done to support multiple single-channel
cameras, all sending packets with the same default VC ID = 0.

Now we check which (and how many) channels are under use on any incoming
FPDLink port, and map it to available channels on the outgoing CSI port.

Signed-off-by: Jai Luthra <me@jailuthra.in>
Signed-off-by: Rishikesh Donadkar <r-donadkar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>