HACK: arm64: dts: ti: k3-am62*-sk: Remove HDMI for Fastboot
commit a6525e3778acc5e209c5b2a42a3783b405b72453 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In the fastboot scenario where display has to come up quickly
on the OLDI panel, the sii902 bridge driver is not made
available. This leads to the probe deferral of the TIDSS
driver and eventually, failure.

This overlay removes the HDMI support for TI K3 AM62* SoCs

Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Aashvij Shenai <a-shenai@ti.com>
Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>