bus: fsl-mc: wait for the MC firmware to complete its boot
commit  add61af3d62113612f74b35b279df9af967f09e7 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

There are use cases in which the Management Complex firmware boot
process is started by the bootloader which does not wait for the boot to
complete. This is mainly done in order to reduce the overall boot time
of a DPAA2 based SoC.

In this kind of circumstance, the fsl-mc bus driver needs to wait for
the MC firmware boot stage to be done before proceeding to the usual
operations such as interrogating the firmware to gather all existent
DPAA2 objects, creating the fsl-mc devices on the bus etc.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>