HSD#15017576547: arm64:dts: agilex: Disable sdm interrupt support
commit fab2c249ff2eed118211e1fce13c65fb8758a6a5 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

fcs in agilex uses svc V1 and sends a poll service status
command to get the response for random number. Before
polling, an interrupt is triggered and the response buffer
is not getting properly populated in ATF when svc
performs service complete smc call.

Disabling an interrupt so the response buffer is properly
populated by the atf when poll service status command is
issued by fcs. Interrupt will be enabled once the
svc client, FCS, moves to svc V3

Signed-off-by: Balsundar Ponnusamy <balsundar.ponnusamy@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>