v4l: xilinx: xcsi2rxss: Fix kernel trace warning
commit 58eb082cd5b933af6db553fad8723ba26f2f5f4e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

As per kernel documentation for call_s_stream, s_stream call should
not be used to start or stop an already started or stopped subdev.
This results in to kernel trace warning on console. To overcome this
remove duplicate calls of s_stream from the start and stop function in
the driver.

Signed-off-by: Kunal Rane <kunal.rane@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
Message-ID: <20250329101554.2079808-1-kunal.rane@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>