AIR-12820-2: neutron: Force to allocate contiguous dma memory
commit  b9102af7e65a1d62c7ae4a88180cce1b81a69249 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The Neutron NPU's SMMU is enabled by the System Manager, but it conflicts
with cache synchronization which is helpful for improving performance.
As a workaround, this patch is used to force the driver to allocate
contiguous DMA memory to avoid issue when doing cache synchronization.

Signed-off-by: Forrest Shi <xuelin.shi@nxp.com>
Signed-off-by: Jiwei.Fu <jiwei.fu@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>