xilinx-ai-engine: Add support for user_event1 (inference complete)
commit 7baefae1845ad72409e58dcd6bc2e34ea60ba036 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Uc module in aie2ps communicates inference complete via user_event1 in
shim tile. This event complete needs to be propagated to aiarm driver.

For user_event1, npi_irq[1] and [2] are used. i.e interrupt 2 and 4.
Uc module in any column can generate the event. Except for col 1,
user_event1 is broadcast through broadcast channel 2. Once it reaches
col1. L2 controller in col 1 generates an interrupt.

Driver does not know which col generated the event. So it has to scan
all the columns to check the status and clear the event.

If aiarm driver registered call back for inference complete, call the cb
function.

In case there is an error event and l1 is backtracked, backtrack
user_event1 as well.

Signed-off-by: Govindarajulu Varadarajan <govindarajulu.varadarajan@amd.com>
Acked-by: Bill Lin <bill.lin@amd.com>
Message-ID: <20250402194204.1402233-42-govindarajulu.varadarajan@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>