staging: xilinx-tsn: Handle inactive MCDMA queues in QBV
commit 0d8a07cb114afaf3a1830fb207398a43ba11cfcc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

When an MCDMA queue deactivated in the QBV configuration receives
frames, MCDMA waits indefinitely for the TSN IP to process that queue,
blocking the handling of other queues. To prevent this problem, disable
the MCDMA queue while it is inactive in the QBV configuration and
re-enable it once the QBV configuration is removed.

Signed-off-by: Nagadheeraj Rottela <nagadheeraj.rottela@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
Message-ID: <20250403085941.2683813-11-nagadheeraj.rottela@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>