Revert "fbdev: Make registered_fb[] private to fbmem.c"
This reverts commit 5727dcfd8486399c40e39d2c08fe36fedab29d99.

There is a issue with seamless display not working, where the
handoff from the initial kernel frame-buffer driver such as simplefb or
simple-drm to the main NVIDIA display driver does not work. This is due
to the kernel function 'num_registered_fb' no longer being exported.
This can be workaround by ...

1. Disabling display handoff in UEFI so that simplefb/simpledrm is not
   used and just have the main display driver start the display.
2. Reverting the kernel change to ensure 'num_registered_fb' is
   exported.

There are more details in this forum thread:
https://forums.developer.nvidia.com/t/seamless-display-handoff-on-bring-your-own-kernel/328849

Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>