remoteproc: xlnx: avoid mailbox setup
commit f0ded4117520e08e1237f3c430a4dcc92d896486 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Mailbox properties are optional in remoteproc xlnx bindings.
If mailbox properties are not found in device-tree it's not fatal error
in the driver. Driver will print warning messages and continue with
normal operation. However, these warning messages can be interpreted as
error. Check "mbox-names" property in driver and setup mailbox only if
it's available in the device-tree. This will avoid any unwanted
warning/error messages.

Signed-off-by: Tanmay Shah <tanmay.shah@amd.com>
Reviewed-by: Ben Levinsky <ben.levinsky@amd.com>
Message-ID: <20250408220556.42291-4-tanmay.shah@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>