PCI: xilinx-cpm: Add support for PCIe RP PERST# signal
commit 304b2d47feb60bbbea5f99200b54252fed772ea5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The "Power Up" sequence mentioned in section 2.2.1 of PCIe
Electromechanical Spec is handled in hardware and enabled during
initialization. With this patch, the reset after the Power Up
sequence will be handled by the driver, which will be useful during
warm reset where the power of the system is already stable.

Add PCIe IP reset along with GPIO-based control for the PCIe Root
Port (RP) PERST# signal.

This change improves the reliability of the reset process by ensuring
that both the Root Port controller and the Endpoint are reset
synchronously, which is critical for maintaining the LTSSM (Link
Training and Status State Machine) state of both components and
avoiding lane errors.

Adapt the implementation to use the GPIO framework for reset signal
handling and make this reset handling optional, along with the
`cpm_crx` property, to maintain backward compatibility with existing
device tree binaries (DTBs).

Signed-off-by: Sai Krishna Musham <sai.krishna.musham@amd.com>
Message-ID: <20250409122054.2432223-3-sai.krishna.musham@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>