net: axienet: Check if auto negotiation is enabled before using it
commit a409bb6d6a471850a7ead81cc1dde56e249ae796 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

XXV PCS feature implementation currently relies on AN status register.
But this feature is optional and inaccessible when AN/LT is disabled in
the IP at the synthesis time. Use the new auto negotiation property to
check for this feature. Default to AN off when the property is not
present to ensure backward compatibility for existing users without AN
in their designs.

Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
Message-ID: <20250411114053.1240266-3-abin.joseph@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>