drm: xlnx: hdmi: Fix random no link/video lock issue
commit 7e092d63a7780efe86a3645724b449a7b9440a58 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

For each display mode, configure the clock on-board clock providers
and redriver chips, which assert the link and video ready signals.
This process takes some time to elevate the video/link ready signals.
The current delay in the driver is insufficient, causing timeouts and
link/video lock issues.

Note: This is delay may need to be reduced in future with proper fix.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
Message-ID: <1744608226-1187105-2-git-send-email-rajesh.gugulothu@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>