fixup! mmc: sdhci-of-dwcmshc: define sdio timeout clocks
commit faa9140ce1086fe2fe86bb6008907d2b62436d6f from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The usage of the .data field of the of_device_id array changed with the
move to kernel 6.11, but because it holds a void pointer there were
no warnings that the sdhci_dwcmshc_rp1_data was now the wrong type.

Convert sdhci_dwcmshc_rp1_data to a dwcmshc_pltfm_data to avoid a crash
when one of RP1's SDIO interfaces is enabled.

See: https://github.com/raspberrypi/linux/issues/6778

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>