extcon: ptn5150: Support USB role switch via connector fwnode
commit  bc488686887d62d7660a342597c00ca661b65f87 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Since the PTN5150 is a Type-C chip, it's common to describe related
properties under the connector node. To align with this, the port
node will be located under the connector node in the future.

To support this layout, retrieve the USB role switch using the
connector's fwnode. For compatibility with existing device trees,
keep the usb_role_switch_get() function.

Acked-by: Chanwoo Choi <cw00.choi@samsung.com>
Reviewed-by: Frank Li <Frank.Li@nxp.com>
Signed-off-by: Xu Yang <xu.yang_2@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>