xlnx: zynqmp_kms: Update planes asynchronously in the legacy entry
commit 9852ada83e40a1a0f142a2375715056f83a8b317 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

With the atomic modesetting, the legacy APIs create a commit for single
change. This serializes each changes with vsync interval as each commit
is synchronous to vsync. So, if application is changing both crtc and
plane within single vsync interval, only one will be taken in a single
interval, and it gives half of the framerate.

This patch makes the plane update asynchronous to vsync, meaning the plane
update from the legacy set plane API happens immediately when there's
no atomic commit queued for the given plane. The implementation is based
on drm_atomic_helper_update_plane() but enables the async update.

Signed-off-by: Rohit Visavalia <rohit.visavalia@amd.com>
Reviewed-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Message-ID: <20250424060751.1564928-1-rohit.visavalia@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>