net: axienet: Fix kernel crash for XXV ethernet
commit e811e7b099e132925d3d6f07bce0faf55a3fc5fb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Kernel crash is observed for the designs which have auto negotiation
disabled because the execution is not exiting from the XXV part of PCS
config function as required. Without this an unintended AXI 1G code is
executed leading to a kernel trace[1]. Fix by returning a default of 0
(no errors) if AN is not present in XXV designs, so that link can be
established as usual.
[1]
<snip>
Call trace:
[    9.123981]  phylink_mii_c22_pcs_config+0x30/0x94
[    9.128677]  axienet_pcs_config+0x90/0x2f4
[    9.132765]  phylink_major_config+0x174/0x42c
[    9.137114]  phylink_mac_initial_config.constprop.0+0xb0/0x120
[    9.142938]  phylink_start+0x8c/0x214
[    9.146593]  axienet_open+0x3b0/0x157c
[    9.150334]  __dev_open+0x100/0x1b4
[    9.153815]  __dev_change_flags+0x1f4/0x2c4
<snip>

Fixes: ce9df069c9d8 ("net: axienet: Check if auto negotiation is enabled before using it")
Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Reviewed-by: Suraj Gupta <suraj.gupta2@amd.com>
Message-ID: <20250425070900.1234237-1-abin.joseph@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>