LF-14857-1 net: enetc: Only set FRM_LEN to the first BD of Tx frame
commit  90b35d88466b2db60795a82f2dba5c69b6806185 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

According to the NETC Block Guide of LS1028A, it says "The shaded fields
in the descriptors indicate that the field is only applicable in the
first BD". And FRM_LEN is one of these shaded fields, so FRM_LEN is
invalid for non-initial BDs.

According to the NETC Block Guide of i.MX95, only the first transmit BD
has FRM_LEN field, the chained transmit BD does not have this field. So
for i.MX95, the situation is the same as for LS1028A. Therefore, remove
the setting of FRM_LEN of non-initial BDs.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>