net: eth: altera: Remove leading garbage in Makefile
A leading '+' sneaked in the Altera Makefile with commit 4460e60ce77d
("HSD#15013542017: net: ethernet: add support for Intel FPGA E-tile
Ethernet driver"). This issue remained undetected for an extended
period because the Altera Ethernet driver is not enabled in our BSP.
The prblem surfaced during a 'make distclean' when I was attempting
to clean my build direcorty.

Removes this leading garbage to fix the Makefile syntax.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>